personal-finance-chatbot/backend/requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
transformers
torch